feat: add a button to create a board in the header
