+++ src/components/protected-route.tsx
@@ -1,5 +1,4 @@
 import { Navigate } from 'react-router-dom';
-import { useAppSelector } from '../app/hooks';
 
 interface ProtectedRouteProps {
   children: JSX.Element;
@@ -7,7 +6,6 @@ interface ProtectedRouteProps {
 
 export const ProtectedRoute = ({ children }: ProtectedRouteProps) => {
   const userId = localStorage.getItem('userId');
-  const Id = useAppSelector((state) => state.auth.userData._id);
 
   if (!userId) {
     return <Navigate to="/" />;
